Rename `mui-lab` to `mui-x`
