+++ packages/formik-mui-x/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "formik-mui-lab",
+  "name": "formik-mui-x",
   "version": "1.0.0-alpha.3",
   "license": "MIT",
   "typings": "dist/main.d.ts",
