+++ packages/formik-mui-x/src/StaticDatePicker.tsx
