+++ packages/formik-mui-x/src/__tests__/StaticDatePicker.test.tsx
