+++ packages/formik-mui-x/src/__tests__/StaticDateTimePicker.test.tsx
