+++ packages/formik-mui/README.md
@@ -1,7 +1,7 @@
 # Formik MUI
 
 ![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui/badge.svg)
-![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui-lab/badge.svg)
+![](https://github.com/stackworx/formik-mui/workflows/Build%20formik-mui-x/badge.svg)
 [![license](https://badgen.now.sh/badge/license/MIT)](./LICENSE)
 
 Bindings for using [Formik](https://github.com/jaredpalmer/formik) with [MUI](https://mui.com/).
@@ -14,7 +14,7 @@ Bindings for using [Formik](https://github.com/jaredpalmer/formik) with [MUI](ht
 MUI >=5
 
 ```bash
-yarn add formik-mui formik-mui-lab
+yarn add formik-mui formik-mui-x
 ```
 
 MUI >=3
