+++ stories/x/DateTimePicker.story.tsx
@@ -4,11 +4,11 @@ import { action } from '@storybook/addon-actions';
 import { Field, Form, Formik } from 'formik';
 import * as React from 'react';
 import * as yup from 'yup';
-import { DateTimePicker } from '../packages/formik-mui-lab/src/DateTimePicker';
-import { DesktopDateTimePicker } from '../packages/formik-mui-lab/src/DesktopDateTimePicker';
-import { MobileDateTimePicker } from '../packages/formik-mui-lab/src/MobileDateTimePicker';
-import FormValues from './FormValues';
-import Wrapper from './Wrapper';
+import { DateTimePicker } from '../../packages/formik-mui-x/src/DateTimePicker';
+import { DesktopDateTimePicker } from '../../packages/formik-mui-x/src/DesktopDateTimePicker';
+import { MobileDateTimePicker } from '../../packages/formik-mui-x/src/MobileDateTimePicker';
+import FormValues from '../FormValues';
+import Wrapper from '../Wrapper';
 
 interface Values {
   date: Date | null;
