+++ stories/x/TimePicker.story.tsx
@@ -4,12 +4,12 @@ import { action } from '@storybook/addon-actions';
 import { Field, Form, Formik } from 'formik';
 import * as React from 'react';
 import * as yup from 'yup';
-import { DesktopTimePicker } from '../packages/formik-mui-lab/src/DesktopTimePicker';
-import { MobileTimePicker } from '../packages/formik-mui-lab/src/MobileTimePicker';
-import { StaticTimePicker } from '../packages/formik-mui-lab/src/StaticTimePicker';
-import { TimePicker } from '../packages/formik-mui-lab/src/TimePicker';
-import FormValues from './FormValues';
-import Wrapper from './Wrapper';
+import { DesktopTimePicker } from '../../packages/formik-mui-x/src/DesktopTimePicker';
+import { MobileTimePicker } from '../../packages/formik-mui-x/src/MobileTimePicker';
+import { StaticTimePicker } from '../../packages/formik-mui-x/src/StaticTimePicker';
+import { TimePicker } from '../../packages/formik-mui-x/src/TimePicker';
+import FormValues from '../FormValues';
+import Wrapper from '../Wrapper';
 
 interface Values {
   time: Date | null;
